+++ IsekaiMod/Content/Features/IsekaiProtagonist/SpecialPower/Excalibur.cs
@@ -1,16 +1,22 @@
 ﻿using IsekaiMod.Utilities;
 using Kingmaker.Blueprints;
 using Kingmaker.Blueprints.Classes;
+using Kingmaker.Blueprints.Classes.Prerequisites;
 using Kingmaker.Blueprints.Classes.Spells;
 using Kingmaker.Blueprints.Items.Ecnchantments;
+using Kingmaker.EntitySystem.Stats;
+using Kingmaker.Enums;
 using Kingmaker.Localization;
+using Kingmaker.RuleSystem;
 using Kingmaker.UI.GenericSlot;
 using Kingmaker.UnitLogic.Abilities;
 using Kingmaker.UnitLogic.Abilities.Blueprints;
 using Kingmaker.UnitLogic.Abilities.Components;
 using Kingmaker.UnitLogic.Commands.Base;
 using Kingmaker.UnitLogic.FactLogic;
+using Kingmaker.UnitLogic.Mechanics;
 using Kingmaker.UnitLogic.Mechanics.Actions;
+using Kingmaker.UnitLogic.Mechanics.Components;
 using Kingmaker.Visual.Animation.Kingmaker.Actions;
 using TabletopTweaks.Core.Utilities;
 using UnityEngine;
@@ -19,8 +25,8 @@ using static IsekaiMod.Main;
 namespace IsekaiMod.Content.Features.IsekaiProtagonist.SpecialPower {
     internal class Excalibur {
         private const string ExcaliburName = "Excalibur";
-        private const string ExcaliburDescBuff = ".............."; // TODO: finish this
-        private static readonly LocalizedString ExcaliburDesc = Helpers.CreateString(IsekaiContext, "Excalibur.Description", "..........."); // TODO: finish this
+        private static readonly LocalizedString ExcaliburDesc = Helpers.CreateString(IsekaiContext, "Excalibur.Description",
+            "Your primary weapon gains the holy and radiant enchantments. Your attack range is increased by 40 feet.");
 
         private static readonly Sprite Icon_Excalibur = AssetLoader.LoadInternal(IsekaiContext, "Features", "ICON_EXCALIBUR.png");
         public static void Add() {
@@ -30,7 +36,7 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.SpecialPower {
 
             var ExcaliburBuff = TTCoreExtensions.CreateBuff("ExcaliburBuff", bp => {
                 bp.SetName(IsekaiContext, ExcaliburName);
-                bp.SetDescription(IsekaiContext, ExcaliburDescBuff);
+                bp.SetDescription(ExcaliburDesc);
                 bp.m_Icon = Icon_Excalibur;
                 bp.AddComponent<BuffEnchantAnyWeapon>(c => {
                     c.m_EnchantmentBlueprint = RadiantEnchantment.ToReference<BlueprintItemEnchantmentReference>();
@@ -39,7 +45,12 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.SpecialPower {
                 bp.AddComponent<BuffEnchantAnyWeapon>(c => {
                     c.m_EnchantmentBlueprint = HolyEnchantment.ToReference<BlueprintItemEnchantmentReference>();
                     c.Slot = EquipSlotBase.SlotType.PrimaryHand;
-                }); // TODO: decide on effects, find way to visual increase weapon size
+                });
+                bp.AddComponent<AddStatBonus>(c => {
+                    c.Descriptor = ModifierDescriptor.UntypedStackable;
+                    c.Stat = StatType.Reach;
+                    c.Value = 40;
+                });
             });
             var ExcaliburAbility = Helpers.CreateBlueprint<BlueprintAbility>(IsekaiContext, "ExcaliburAbility", bp => {
                 bp.SetName(IsekaiContext, ExcaliburName);
@@ -47,12 +58,21 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.SpecialPower {
                 bp.AddComponent<AbilityEffectRunAction>(c => {
                     c.Actions = ActionFlow.DoSingle<ContextActionApplyBuff>(c => {
                         c.m_Buff = ExcaliburBuff.ToReference<BlueprintBuffReference>();
-                        c.UseDurationSeconds = true;
-                        c.DurationSeconds = 600;
+                        c.DurationValue = new ContextDurationValue() {
+                            Rate = DurationRate.Minutes,
+                            DiceType = DiceType.Zero,
+                            DiceCountValue = 0,
+                            BonusValue = Values.CreateContextRankValue(AbilityRankType.Default),
+                            m_IsExtendable = true,
+                        };
                     });
                 });
                 bp.AddComponent<SpellComponent>(c => {
                     c.School = SpellSchool.Enchantment;
+                });
+                bp.AddComponent<ContextRankConfig>(c => {
+                    c.m_Type = AbilityRankType.Default;
+                    c.m_BaseValueType = ContextRankBaseValueType.CharacterLevel;
                 });
                 bp.m_Icon = Icon_Excalibur;
                 bp.Type = AbilityType.Supernatural;
@@ -64,7 +84,7 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.SpecialPower {
                 bp.Animation = UnitAnimationActionCastSpell.CastAnimationStyle.Self;
                 bp.ActionType = UnitCommand.CommandType.Standard;
                 bp.AvailableMetamagic = Metamagic.Quicken;
-                bp.LocalizedDuration = StaticReferences.Strings.Duration.OneDay;
+                bp.LocalizedDuration = StaticReferences.Strings.Duration.OneMinutePerLevel;
                 bp.LocalizedSavingThrow = StaticReferences.Strings.Null;
             });
             var ExcaliburFeature = Helpers.CreateBlueprint<BlueprintFeature>(IsekaiContext, "ExcaliburFeature", bp => {
@@ -75,6 +95,10 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.SpecialPower {
                     c.m_Facts = new BlueprintUnitFactReference[] {
                         ExcaliburAbility.ToReference<BlueprintUnitFactReference>()
                     };
+                });
+                bp.AddComponent<PrerequisiteCharacterLevel>(c => {
+                    c.Group = Prerequisite.GroupType.Any;
+                    c.Level = 10;
                 });
             });